requirements.txt
fastapi
uvicorn
pyyaml
python-multipart
fastapi
seaborn